requirements.txt
basicauth==0.4.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
boto3==1.12.34
botocore==1.15.34
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
ckanapi==4.3
click==7.1.1
colorama==0.4.3
colorlog==4.1.0
cryptography==2.8
dash==1.10.0
dash-bootstrap-components==0.9.1
dash-core-components==1.9.0
dash-html-components==1.0.3
dash-renderer==1.3.0
dash-table==4.6.2
decorator==4.4.2
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
email-validator==1.0.5
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
gunicorn==20.0.4
hdx-python-api==4.3.5
hdx-python-country==2.4.8
hdx-python-utilities==2.2.4
html5lib==1.0.1
idna==2.9
ijson==2.6.1
importlib-metadata==1.6.0
inflect==4.1.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
jmespath==0.9.5
jsonlines==1.2.0
jsonpath-rw==1.4.0
libhxl==4.18
linear-tsv==1.1.0
MarkupSafe==1.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
ndg-httpsclient==0.5.1
num2words==0.5.10
numpy==1.18.1
openpyxl==3.0.3
pandas==0.25.0
paramiko==2.7.1
plotly==4.6.0
ply==3.11
psycopg2-binary==2.8.4
pyaml==20.3.1
pyasn1==0.4.8
pycparser==2.20
PyNaCl==1.3.0
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-io-wrapper==0.1
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.3.1
quantulum3==0.7.3
ratelimit==2.2.1
requests==2.23.0
retrying==1.3.3
s3transfer==0.3.3
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.15
sshtunnel==0.1.5
tabulator==1.38.1
text-unidecode==1.3
typing==3.7.4.1
unicodecsv==0.14.1
Unidecode==1.1.1
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
xlrd==1.2.0
yamlloader==0.5.5
zipp==3.1.0
